added append() to Wrapped/Share as proxy to Native/Share
