+++ tests/NativeShareTest.php
@@ -39,19 +39,4 @@ class NativeShareTest extends AbstractShareTest {
 		}
 		$this->share->mkdir($this->root);
 	}
-
-    public function testAppendStream() {
-        $fh = $this->share->append($this->root . '/' . $name);
-        fwrite($fh, 'foo');
-        fclose($fh);
-
-        $fh = $this->share->append($this->root . '/' . $name);
-        fwrite($fh, 'bar');
-        fclose($fh);
-
-        $tmpFile1 = tempnam('/tmp', 'smb_test_');
-        $this->assertEquals('foobar', file_get_contents($tmpFile1));
-        $this->share->del($this->root . '/' . $name);
-        unlink($tmpFile1);
-    }
 }
